be dumb when splitting cli args and smart when creating options
`ArgSplitter` now only splits according to the goal before it on the command
line, without trying to split up the options by scope itself. Then, when
`OptionsBootstrapper` calls `Options.create()`, register all the arguments, then
further parse the cli args with the knowledge of all the registered arguments.

This allows us to use `--binaries-baseurls` on the command line, which was
previously parsed as `{'binaries': ['--baseurls']}`, which no longer
exists.