Create lint-v2 rule and implement it for Python
This is a prerequisite for ultimately adding a pre-commit hook that
behaves similarly to our isort pre-commit hook (warns the user and lets
them actually modify their code once they are ready for it)

* Create a top level `lint-v2` rule;
* Implement `lint_with_black` which implements that rule for python;
* Deduplicate `lint_with_black` with `fmt_with_black`;
* Make use of `FallibleExecuteProcessRequest` to make lint output less
  spammy.

Rename FmtTarget to TargetWithSources

Now we introduced a lint goal which takes the same kind of targets as an
input, it makes sense to use a more general name.

Use UnionMembership to reduce coupling in fmt and lint v2 rules

Improve comment in fmt and lint rules

The issue that was link before isn't obviously related to the issue.
The actual problem is `TargetAdaptor.sources` throwing an exception.

Make two functions private in python_fmt